Moving tests that use MessageStreamAssert to same package name in test folder to use package private
