+++ samza-test/src/test/java/org/apache/samza/test/framework/BroadcastAssertApp.java
@@ -17,14 +17,13 @@
  * under the License.
  */
 
-package org.apache.samza.test.operator;
+package org.apache.samza.test.framework;
 
 import org.apache.samza.application.StreamApplication;
 import org.apache.samza.config.Config;
 import org.apache.samza.operators.MessageStream;
 import org.apache.samza.operators.StreamGraph;
 import org.apache.samza.serializers.JsonSerdeV2;
-import org.apache.samza.test.framework.MessageStreamAssert;
 import org.apache.samza.test.operator.data.PageView;
 
 import java.util.Arrays;
