+++ samza-test/src/test/java/org/apache/samza/test/framework/TimerTest.java
@@ -17,14 +17,14 @@
  * under the License.
  */
 
-package org.apache.samza.test.timer;
+package org.apache.samza.test.framework;
 
 import org.apache.samza.test.operator.StreamApplicationIntegrationTestHarness;
 import org.junit.Before;
 import org.junit.Test;
 
 
-import static org.apache.samza.test.timer.TestTimerApp.PAGE_VIEWS;
+import static org.apache.samza.test.framework.TestTimerApp.PAGE_VIEWS;
 
 public class TimerTest extends StreamApplicationIntegrationTestHarness {
